Fix While Clause Being Checked Inside of Lines
**Bug**
In VS Code in the markdown grammar, we see some grammar rules that use `while` stop matching early on the second line of the rule (the line after the begin). The root cause is that the while check clause is being evaluated on both on the start of the second line, and after the first match on the second line.

**Fix**
The while expression should only be checked on lines to determine if the line should be matched or not. Once a line has been matched with while, the entire line becomes part of the group. This fix adds a second condition to the while check so that we only apply it when starting a line.

I also added a simple grammar to test this behavior. It failed to match properly before, but does work properly after the fix.